Change shaft_element.py line endings.
The file `shaft_element.py` had the end of lines with CRLF. The standard for
cross-platform development is to let git replace those with LF.

One way to control this is to use the .gitattributes file with *
text=auto. See #18.

Another recomendation is for windows developers to use the following
config:
$ git config --global core.autocrlf true

More info can be found here
https://git-scm.com/book/en/v2/Customizing-Git-Git-Configuration#Formatting-and-Whitespace