Remove plot when creating rotor
The rotor geometry was plotted when the object was instantiated. The
purpose of this was to show the user the rotor geometry so that the user
could check if everything was correct before eigenvalues, eigenvectors
etc. were calculated.
The problem with this approach is that the rotor will be automatically
plotted in situations where this is not wanted by the user, such as when
running a script where several rotors will be instantiated or when we
are running the tests locally, in this case, several windows with plots
will appear and the tests will take a long time.
With this PR the plot of the rotor after instantiation will be optional
to the user. A message is provided to let the user know that the
plot_rotor() method can be used.
Additionally, the message "calculate eigenvalues and frequencies", which
is redundant, has been modified to "calculate rotor state".