Remove imports from __init__ to simplify interface
This removes imports such as rst and bse, since the user will be able
to access those with rs.bearing_seal_element, and if the user wants to
use the BearingElement class for example, it can be done with
rs.BearingElement, since there is already a line with
'from .bearing_seal_element import *'.
For tests this changes from ross import * to import ross as rs.
Although it is not considered a bad practice to use * on the test files
we should try to replicate on our tests how the user will call the
library.