Change appveyor badge to track master
This changes the appveyor badge that appears in the repository main page
to track only builds that are related to the master branch.
We should have the badge only for the master branch since PRs with
failing builds will not be merged anyway.
The previous badge was tracking all builds, including those from new
pull requests that were failing before merging to master.