Add import to Axes3d
The import has been added to the results module because the mode shape
plot needs this to work. Although it looks like the import is not being
used (this is highlighted by linters), matplotlib, for some strange
reason, actually needs this to use the projection='3d'.
A doctest has been added to the mode shape plot for the tests to fail
if this import is removed.