Add xlrd exception
There was an error when trying to resolve a xlrd exception, in which the
sheet_name argument was not being recognized. This argument is not part
of the read_csv function, as a csv file holds no more than one sheet.
The xlrd exception was included again, except this time it will no
longer use the sheet_name value.