Drop NaNs from table
The chosen strategy to drop NaNs begins by droping columns in which all
values are NaNs, then drops rows in which at least one value is NaN.
Depending on the table, important rows could be droped this way (let us
say, if there is at least one cell that is not a NaN in a column filled
mostly with NaNs. Still, we are now able to tackle some problems we were
having.