Add bearing location to DataFrame
This adds the bearing location to the DataFrame. To define the axial
location there is a loop that goes through the bearings checking there
nodes and setting the axial position (nodes_pos) accordingly. To define
the y position, bearings that are on the same axial position are
grouped together and the y position is set from the bearing connected
to the shaft to the bearing connected to the ground. This assumes that
the bearings are connected with nodes that have increasing order from
the shaft to the ground.
As an example:
bearing = rs.BearingElement(0, n_link=7, kxx=stfx, kyy=stfy, cxx=0)
bearing_house = rs.BearingElement(7, n_link=8, kxx=kxx, kyy=kyy, cxx=0)
support = rs.BearingElement(8, kxx=stfx, kyy=stfy, cxx=0)