Fix matplot_eccentricity doctests
The problem was the following:
The last plot example before fluid_flow_graphics was this:

    Example
    -------
    >>> rotor = rotor_example()
    >>> modal = rotor.run_modal(speed=0)
    >>> modal.wn[:2]
    array([91.79655318, 96.28899977])
    >>> modal.wd[:2]
    array([91.79655318, 96.28899977])
    >>> modal.plot_mode(0) # doctest: +ELLIPSIS
    (<Figure ...

This one uses the ax.plot(projection='3d'). When we started the
matplot_eccentricity test we go through these lines:

if ax is None:
    ax = plt.gca() # (get the current axes)

Since we were not providing an axes in the example, the code would try
to use the current axes which had a 3d projection and this would cause
the error reported above.

We can eliminate this problem by changing the example to:

    >>> import matplotlib.pyplot as plt
    >>> from ross.fluid_flow.fluid_flow import pressure_matrix_example
    >>> my_fluid_flow = pressure_matrix_example()
    >>> fig, ax = plt.subplots()
    >>> ax = matplot_eccentricity(my_fluid_flow, z=int(my_fluid_flow.nz/2), ax=ax)
    >>> # to show the plots you can use:
    >>> # plt.show()