Remove docstring example
I am removing this docstring example for now. The problem here is that
this test creates a file when we run it, and this is not good since it
will add not needed files to the project folder. This is a problem that
we also have for other tests and we need to try and find a way to solve
this, until then I think it is better to remove the example/test.