Add synchronous option to plot_ucs
This adds an option to run plot_ucs with synchronous option, where the
rotor speed will be equal to the frequency of the first forward mode.
This was implemented with the following steps:
1. Run modal analysis with speed=0;
2. Get frequency of first forward mode;
3. Run modal analysis with speed=<frequency from step 2>.

The time to run the speed=0 method:
10.3 s ± 1.03 s per loop (mean ± std. dev. of 7 runs, 1 loop each)

The time to run synchronous=True method:
42.7 s ± 3.24 s per loop (mean ± std. dev. of 7 runs, 1 loop each)

Regarding this method:

"Also, the rotor transverse inertia, It, is replaced with the
transverse minus the polar inertia (It-Ip). This last modification, in
effect, constrains the rotor rotational speed to equal the eigenvalue
frequency."

We are still evaluating how to implement it, but it should give a good
performance improvement.