Fix isinstance check
This checks if the element is not an instance of SealElement, since the
ucs is carried out only with the bearings. The code before compared the
name of the class with 'BearingElement', the problem with that is that
if we have a bearing that inherits from 'BearingElement' the name of
the class will not be 'BearingElement' but we have to consider it in
the calculation.