Add script to generate plots for comparison
This adds a script that generates some plots for 3 different rotor
models. This helps to solve a problem that we had when we modified some
property to improve the plot for a specific model and later we would
notice that at the end the modification would be detrimental to other
models.
This, together with the plotly_theme style, closes #580.
Now we should try to make all plot modifications on the plotly_theme.py
file, and we should run the generate_comparison_plots.py script and
check the output.
After running the script we can open the index.html file and click each
link to check the specific plot for the three models.
This script has been added to a tools directory inside the repository.
The tools directory will be used to store some utility used for the
development of our code. This is the same structure that other projects
such as numpy use.
I have added the generated html files to the .gitignore files, since
they should always be generated by the script (with the exception of
the index.html file).