Modify available_materials.toml according to save
When running the ross tutorial, an example where we would loaded the
steel from available_materials.toml and then would saved it (with
steel.save_material()) would result in a modification of the
available_materials.toml file. What I have done here was to load all
materials that we have in the file for now and then saved each material.
Now if we do this operation the file won't be changed.