Add option for scale_factor='mass' in disks
This adds the option to use scale_factor='mass' when creating the disks.
If all the disks given to the rotor constructor have this scale factor,
the value for the scale factor will be calculated when the rotor is
instantiated based on the disk with the highest mass.
The disks with mass close to the highest value will have scale_factor=1
and disks with lower masses will have scale factors that will be closer
to 0.5.