Fix rotor_example_6dof docstring
This fix comments in the code such as:

    Plotting rotor model

becomes:

    >>> # Plotting rotor model

to avoid errors when running the tests.

Calls to fig.show() have been commented out to avoid having to provide
and output repr for comparison.
The line break (\n) within strings have been removed, since this would
cause errors as well.