Add .get_application_registry to pint import
With pint we cannot have two quantities from different registries
interacting. This can lead to problems when we have an application that
uses pint calling a package that also imports pint, since the
application registry will be different from the package registry.

To solve this problem we need to use pint.get_application_registry
within our package, so that the package (ross in this case) can use the
same registry as the application.

One problem that occured after using .get_application_registry is that
it was not possible to use the .load_definitions method, since this
would cause a unit 'redefinition' which is not allowed by pint.

To solve this, after calling .get_application_registry, we check if the
object is an instance of LazyRegistry. Basically, what this means is
that .set_application_registry has not been called by an application, so
we can create a new registry with .UnitRegistry and .load_definitions.
For this to work, we also had to move the call to pint.Quantity([]). Now
it comes after setting the registry, since calling it before would cause
the default registry to be of the UnitRegistry type instead of
LazyRegistry.

After that we use the .set_application_registry method to make the
quantities pickable, which is required by the multiprocessing module.
The check for ureg instance can change in the future, since the code now
in the pint master branch has merged the ApplicationRegistry wrapper.
In this case we can get the instance of the registry by using:

isinstance(ureg.get(), pint.registry.LazyRegistry)