Fix stiffness expression for tapered element
In the stiffness expression described on appendix A of the Genta and
Gugliotta paper, when substituting chi = 12*E*I/(phi*G*A*L**2) in the
constant that multiplies the second matrix, we had canceled the A here
with the Aj from the equation, but they are actually not the same area.
The area Aj in the equation refers to the area in the element's left
side, while the area in the chi calculation refers to the area in the
middle of the element. Therefore, we were having an error in this
formula by a factor of Aj / A.