Move shape calculation from init to _calculate
The purpose of moving the calculation to a specific method is to avoid
performance issues when running a modal analysis, since we have an
instantiation of a shape object for each mode when we call run_modal.
This would be very expensive when running a campbell diagram for
example. Now we are only going to run the calculation when the user
calls a plot method, and if the calculation has already been performed
(self.orbit is not None), calculation is skipped.