Store modal results inside campbell results
It is common that, after creating a Campbell diagram, the user will
want to explore some modes that are shown in the plot. Since the
calculation for that mode has already been carried out, it makes sense
for us to store that result so that we don't need to run it again.